Former-commit-id: 028c41f87ae938ce1c61363a14bb21d2079cb914

+++ scripts/instagrabber.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 import errno
+import json
 import logging
 import os
 import requests
@@ -17,6 +18,7 @@ output_dir = "assets/instagram_photos"
 
 
 output_dir = os.path.abspath(output_dir)
+big_json = output_dir + ".json"
 
 try:
     os.makedirs(output_dir)
@@ -42,19 +44,51 @@ def get_images():
             pass  # fuck videos man
 
         # Name file after media id
-        filename = media.id + ".jpg"
-        output_file = os.path.join(output_dir, filename)
-        if not os.path.exists(output_file):
-            logging.info("Instagrabber: Downloading {}...".format(filename))
+        output_file = os.path.join(output_dir, media.id + ".jpg")
+        output_json = os.path.join(output_dir, media.id + ".json")
+        if not (os.path.exists(output_file) and os.path.exists(output_json)):
+            logging.info("Instagrabber: Downloading {}...".format(media.id))
             r = requests.get(media.images['standard_resolution'].url)
             with open(output_file, 'wb') as f:
                 f.write(r.content)
 
+            # JSON data
+            file_info = json.dumps(
+                {
+                    # "caption": media.caption,
+                    "filename": output_file,
+                    "tags": [tag.name for tag in media.tags]
+                }
+            )
+            with open(output_json, 'wb') as j:
+                j.write(file_info)
+
+
+def refresh_json():
+    with open(big_json, 'wb') as j:
+        big_data = []
+        for json_file in os.listdir(output_dir):
+            if json_file.endswith(".json"):
+                file_path = os.path.join(output_dir, json_file)
+                try:
+                    json_data = open(file_path).read()
+                    data = json.loads(json_data)
+                    big_data.append(data)
+                except:
+                    pass
+
+        j.write(json.dumps(big_data))
+
+
+def be_righteous():
+    get_images()
+    refresh_json()
+
 
 if __name__ == "__main__":
-    get_images()  # do one now for good measure
+    be_righteous()  # do one now for good measure
     scheduler = BlockingScheduler()
-    scheduler.add_job(get_images, 'interval', minutes=15)
+    scheduler.add_job(be_righteous, 'interval', minutes=15)
 
     try:
         scheduler.start()
